backend/requirements.txt
Django==3.2.25
django-cors-headers==4.3.1
django-postgres-extensions==0.9.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
netifaces==0.11.0
gunicorn==22.0.0
docutils==0.21.2